added difficulty
